Feat/reports

See merge request epm-edai/project-runs/run-20/team-2/serverless/restaurant-app!59

+++ backend/restaurant-backend-app/dnapp/lambdas/restaurant-api/Restaurant.Api/Contracts/Responses/OrderResponse.cs
@@ -15,7 +15,7 @@ public sealed class OrderResponse
     public int GuestsCount { get; set; }
     public string Status { get; set; } = null!;
     public List<OrderDishSnapshot> Dishes { get; set; } = [];
-    public float TotalAmount { get; set; }
+    public decimal TotalAmount { get; set; }
     public string CreatedAt { get; set; } = null!;
     public string? CompletedAt { get; set; }
 }
